Move submission statistics report to /stats
