+++ picoCTF-web/api/apps/v1/submissions.py
@@ -77,19 +77,3 @@ class SubmissionList(Resource):
         return jsonify({
             'success': True
         })
-
-
-@ns.response(200, 'Success')
-@ns.response(401, 'Not logged in')
-@ns.response(403, 'Not authorized')
-@ns.route('/statistics')
-class SubmissionStatistics(Resource):
-    """View submission statistics, broken down by problem."""
-
-    @require_admin
-    def get(self):
-        """Get submission statistics, broken down by problem name."""
-        return jsonify({
-            p['name']: api.stats.get_problem_submission_stats(p['pid'])
-            for p in api.problem.get_all_problems(show_disabled=True)
-        })
